pyproject.toml
[project]
name = "pycanstreamviewer"
version = "0.1.0"
description = "Live CAN bus viewer with PyQt6 and pyqtgraph"
authors = [
    {name = "Victor Zaccardo", email = "victorzaccardo@gmail.com"},
]
requires-python = ">=3.11,<3.15"
license = {text = "MIT"}
dependencies = [
    "PyQt6>=6.7.1",
    "pyqtgraph>=0.13.7",
    "python-can>=4.4.2",
    "cantools>=39.4.11",
    "PyYAML>=6.0",
    "numpy>=1.26.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[project.optional-dependencies]
distribution = [
    "PyInstaller>=6.0",
]

[tool.pdm.scripts]
start = "python -m pycanstreamviewer"
dist = "python -m pycanstreamviewer.distribution"
